Fix bug in debounce code
When misfiring, set the last observed value to the current state of the
sensor in WAIT_FOR_DEBOUNCE, which is now HIGH again. If we don't do
this, and the noise pulse is so close to the stable pulse that the
stable pulse begins on the very next iteration of loop(), then the real
pulse would be missed (because WAIT_FOR_INPUT is looking for a
high-to-low transition). This has only been seen with a VERY noisy
mechanical switch sensor.